refactor: economy phase-2 slice 10 — consumption scenario tables (#3856)
Migrate resolveConsumption and per-phase consumption helper tests to
table-driven scenarios in colonizethis_economy_test_support.
