+++ packages/colonizethis_economy_test_support/lib/colonizethis_economy_test_support.dart
@@ -12,6 +12,8 @@ export 'src/bid_spend_parity_scenarios.dart';
 export 'src/boycott_blocked_commodities_scenarios.dart';
 export 'src/boycott_blocked_commodities_test_support.dart';
 export 'src/carry_forward_bid_notional_scenarios.dart';
+export 'src/consumption_phases_scenarios.dart';
+export 'src/consumption_scenarios.dart';
 export 'src/core_economy_test_support.dart';
 export 'src/deal_matcher_frr_scenarios.dart';
 export 'src/deal_matcher_priority_scenarios.dart';
